Switch to vi-style key bindings for fish
With the fish v2.3.0 release, they've fixed all of my annoyances with vi-mode,
so things like Ctrl-L and Ctrl-F work as expected out of the box. And, if
I should need to change bindings in the future, the normal user key bindins are
honored, so no need for the funky work arounds.