Do not use vi-style key bindings for fish shell
I relied on the hybrid behavior from before fish v2.5, which is now
available as a preset key binding, but the Vi mode has always been a bit
wonky. With fish v2.5, there is now a built-in `edit_command_buffer`
function that's bound to Alt-e and Alt-v out of the box which will drop
you into a real editor when necessary. I'll use that instead and rely on
the normal default key bindings.