Tweak kitty key mappings
The `ctrl+b` was a bad idea because it interfered with Vim. I could not
figure out how to get the clear control character commands to work (it
just always rendered "[2j" on the screen...will tinker more with that.
I also added the tab navigation which is more natural for me than using
the arrow keys.