Add initial Kakoune setup
Using plug.kak to manage plugins (including itself). I'm only storing
a raw copy of the plugin within this repo, so to bootstrap managing
plug.kak with itself on a fresh dotfiles clone:

1. run `:plug-list` and find the plugin;
   it should show "Local changes"
2. clean the plugin by highlighting it and pressing `D`;
   it should show "Deleted"
3. install the plugin again by highlighting it and pressing `I`;
   it should show "Done"

Note that trying to symlink plug.kak within Kakoune's autoload/
directory caused problems with highlighters and other functionality, but
removing the directory fixed things up, so we'll source it explicitly.
The highlighting problem persisted even with an empty autoload/
directory, so it's likely an upstream bug, or I'm not 100% understanding
the functionality.

See: https://github.com/andreyorst/plug.kak