Ensure tridactyl config file changes are exhaustive
Without sanitizing the local cache when sourcing the configuration
file, certain settings could drift over time and leave things in an
unexpected state. Instead, we can force the plugin to start fresh with
the default settings each time the config is sourced so we know that the
customizations accurately reflect all changes.