Remove J and K overrides for tab switching in Gmail
Unfortunately, Tridactyl doesn't seem to pay attention to whether or not
those characters are within an input field or not, so right now I can't
type a capital 'J' or 'K' in an email without switching tabs. That seems
like a bug, or perhaps there's an easy workaround, but for now I'll just
remove the bindings.