Add Ctrl-G binding to fuzzy select local Git branch names
It was easier to modify this file to add the function rather than put it
elsewhere, but it may make pulling down changes from upstream more
difficult if I need to do that in the future. If it becomes problematic,
I'll split it out.

The preview will show the most recent 200 commits on the selected
branch.