Switch to normal Vim keys for i3wm directions
The default `jkl;` is inconsistent with my muscle memory of Vim's
`hjkl`, which is used for everything else in my world. This does create
a conflict with the horizontal split binding, so I've made that `mod+b`
which matches the default that Sway uses. More often than not, I'll use
the `mod+e` toggle binding anyway.