Set i3wm workspaces to stacking by default
I was swayed by the arguments made by /u/Adaephon-R in a thread on
Reddit about consistency in left/right movement only being for switching
containers. I've also removed the status bar from the top since it
covers up the window titles which are useful to understand navigation
(more so than the bottom portion of the screen).

See:
- https://old.reddit.com/r/i3wm/comments/3yp55u/realworld_examples_using_stacked_and_tabbed/cyg2c6q/