Switch kitty font to DejaVu Sans Mono family
This doesn't have the ligatures like Fira Code, but it has an oblique
version for italics, it works at a smaller font size so I can have more
on a screen, and it renders a lot more clearly. It's otherwise pretty
similar looking, and perhaps in the future with more font rendering
tweaks, I could switch back, but for now this seems to work best.