Copy example xss-lock script to lock before sleeping
This is required otherwise the lock will happen after the laptop wakes
up, and there can be a bit of a delay or a time when the desktop shows.
Instead, we want to ensure we lock and then sleep first. This was copied
from `/usr/share/doc/xss-lock/transfer-sleep-lock-i3lock.sh`.