Use system mode for default exit key binding
The custom mode covers all use cases. I've also switched Return to
default to lock to make it a bit easier to hit as the most common
option.

The reload config was removed since a restart is almost always better
and works just fine. I might re-bind the `$mod-c` to something else in
the future.