Allow cancelling the capslock when one shift is pressed
With faster typing, I've sometimes accidentally enabled capslock by
hitting both shift keys at the same time rather than doing it on
purpose. I'm going to try this to see if it's easier to get out of this
way. I so rarely actually want capslock, so it's a much more deliberate
act.