Use spacebar as the i3wm mod key when held down
A lot of instructions for changing keymaps use the older `xmodmap`
command, which is no longer recommended. Instead, generating a copy of
the existing keymap and then modifying it is the way to go:

    $ xkbcomp $DISPLAY ~/.xkeymap

Normal fast typing doesn't seem to be a problem, but I do need to be
precise. We'll see if this causes too many frustrations, but it's much
more comfortable compared to using the windows key (which still works,
as both are considered Mod4).

IMPORTANT: You must remap some other unused key to space or it won't be
available for `xcape` to remap later. I chose <I159> "XF86DOS" since
it's not actually on my physical keyboard and I don't think I'd ever
use that particular mapping for anything else...there's no significant
meaning behind using that particular key.

This also enables tapping the Control key to be interpreted as Escape.
Even though my muscle memory is pretty used to Ctrl+[ for Escape, that
doesn't work everywhere, so this could be handy to have in my back
pocket.

Note that I don't believe the `setxkbmap` command is required anymore
since those options are also defined in the keymap file itself, but I'll
leave it for ease of backing this out.

See:
- https://old.reddit.com/r/i3wm/comments/5zpz69/using_space_bar_as_mod_is_life_changing/
- https://github.com/alols/xcape
- https://wiki.archlinux.org/index.php/Xorg/Keyboard_configuration#One-click_key_functions
- https://wiki.archlinux.org/index.php/X_keyboard_extension#Using_keymap