Force use of kitty-terminal for lauching external commands
This is much faster than what `repl` was doing by launching x11
terminals, which adds lag due to the startup of a fresh instance of
kitty. Forcing the use of `kitty-terminal` instead will use a single
instance and opens things at the bottom of the existing window rather
than making a full window managed by i3.

In order for these commands to work, we must allow remote control of
kitty or the kakoune integration will fail.

See:
- https://github.com/mawww/kakoune/blob/master/rc/windowing/kitty.kak
- https://sw.kovidgoyal.net/kitty/remote-control.html