Make the PrtSc key function as Super as well
The print screen key on the Lenovo T495 is mirrored in position to the
Windows key. That is considered the `$mod` key for i3wm purposes, so now
I can at least use either the left or right side of the keyboard for
commands. Not quite as ergonomic as using the spacebar previously, but
we'll see how it feels with a bit of use.

I also went back to plain `shift:both_capslock` as having the `_cancel`
on there meant you could't type an underscore without leaving the
capslock mode, and then turning it back on (it would just type a
hyphen). I'd much rather be able to type common identifiers in all caps
(like environment variables) and have to remember to disable capslock
afterward.