Switch to chezmoi instead of Stow
I have to install something either way to manage the dotfiles (I don't
want to go back to a custom shell script), and chezmoi supports a bunch
of useful features, has a thriving community, and it's written in Go.

Unfortunately, stow has some rough edge cases and the sole maintainer
has been burnt out for a number of years.

See:
- https://www.chezmoi.io
- https://github.com/aspiers/stow/issues/33