Re-enable default auto-pairs behavior
I had turned this off before as it wasn't a very smooth interaction, but
the recent release of Helix added "smart tab" to help move to the cursor
to after the closing brace, and indentation also seems more solid. Let's
try sticking with the upstream default once more and see if I can get
used to it.

See:
- https://helix-editor.com/news/release-23-10-highlights/