Enable automatic Git commit signing with SSH keys
This works across both Git and jj, and has also been configured on
my mobile devices. It's not a perfect system, as revocation isn't
really supported, and we rely on the provider (GitHub) to determine
verification rather than the web of trust when using GPG, but the
usability is so much better with SSH keys.